requirements.txt
nltk==3.5
numpy==1.19.1
scipy==1.4.1
scikit_learn==0.23.2
transformers==3.0.2
pytorch==1.6.0